% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psa_data}
\alias{psa_data}
\title{Sample PSA data}
\format{
\subsection{\code{psa_data}}{

A list of data frames, one for each scenario corresponding
to the list item name (\code{Baseline}, \code{Intervention_1}). Each data frame has 100
rows and 814 columns:
\describe{
\item{date}{(\code{Baseline} only) the simulation date in calendar time}
\item{time}{(Intervention scenarios only) simulation time as integer
starting with 0}
\item{X1, ..., X1727}{individual simulation results}
}
}
}
\usage{
psa_data
}
\description{
Sample simulation outputs for a baseline scenario and
three interventions, in a format compatible with package
functions
}
\keyword{datasets}
